requirements.txt
# Core ML & Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.4.0

# Timing Model
prophet==1.1.5

# Content Model
anthropic==0.18.1

# API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.3
apify-client==1.6.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# Evaluation
bert-score==0.3.13
